refact: shrink direct-write PR diff for review
Move gRPC mock cluster and topic client stub into topicwritetest,
merge reapply test, consolidate reconnector stream-connect handling,
and simplify partition location conversion.
